+++ packages/components/src/components/drawer/figma/drawer.figma.lite.tsx
@@ -1,5 +1,6 @@
 import { useMetadata } from '@builder.io/mitosis';
-import { DBDrawer } from '../index';
+import DBDrawerHeader from '../../drawer-header/drawer-header.lite';
+import DBDrawer from '../drawer.lite';
 import { FigmaDrawerProps, drawers } from './drawer.figma';
 
 useMetadata({
@@ -7,5 +8,14 @@ useMetadata({
 });
 
 export default function DrawerFigmaLite(props: FigmaDrawerProps) {
-	return <DBDrawer>{props._children}</DBDrawer>;
+	return (
+		<DBDrawer
+			header={
+				<DBDrawerHeader closeButtonText="Close">
+					{props._showText ? props._drawerHeader : ''}
+				</DBDrawerHeader>
+			}>
+			{props._children}
+		</DBDrawer>
+	);
 }