+++ showcases/react-showcase/src/components/form/index.tsx
@@ -6,6 +6,7 @@ import {
 	DBCustomSelect,
 	DBDivider,
 	DBDrawer,
+	DBDrawerHeader,
 	DBInfotext,
 	DBInput,
 	DBLink,
@@ -509,9 +510,9 @@ const FormComponent = () => {
 					onClose={() => {
 						setDrawerOpen(false);
 					}}
-					open={drawerOpen}
-					spacing="none">
-					<DBInfotext style={{ margin: '100px', display: 'flex' }}>
+					header={<DBDrawerHeader>Drawer Test</DBDrawerHeader>}
+					open={drawerOpen}>
+					<DBInfotext>
 						Test: Mouse down here, drag to backdrop, release. Drawer
 						should stay open.
 						<DBTooltip placement="bottom-start">