+++ packages/postcss-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @db-ux/core-postcss-plugin
 
+## 5.0.1
+
+_version bump_
+
+## 5.0.0
+
+_version bump_
+
 ## 4.14.0
 
 _version bump_